Ensure that there are scripts before using its collection
The current code expects the collection of scripts to be populated. However that is not the case when it runs in a JS interpreter that has nothing in the DOM.

It may happen when running ScalaJS tests in Rhino.

Check this thread to see more details: https://groups.google.com/d/msg/mozilla-rhino/Ooir-2R29-w/VRc1CWogXIcJ